Make it closer to a simple move
